Move Options to Configuration Namespace and Folder
