+++ MouseUnSnag/CommandLine/CommandLineParser.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
+using MouseUnSnag.Configuration;
 
 namespace MouseUnSnag.CommandLine
 {
