+++ MouseUnSnag/TrayIconApplicationContext.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 using MouseUnSnag.CommandLine;
+using MouseUnSnag.Configuration;
 
 namespace MouseUnSnag
 {
